refactor: move deep_copy from _init.py to _utils.py
Deep copy is a utility/cloning concern, not initialization.
